% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_bls_data}
\alias{get_bls_data}
\title{Get data from the BLS public API.}
\usage{
get_bls_data(series_ids = "LAUCN040010000000005", registration_key = NA)
}
\arguments{
\item{series_id}{input character vector}
}
\value{
data.frame with data for the requested series id.
}
\description{
Get data from the BLS public API.
}
\examples{
get_bls_data(registration_key='BLS-provided key')
get_bls_data('LAUCN040010000000005', registration_key='BLS-provided key')
}

